requirements.txt
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
openai
python-dotenv
schedule